KiraClaw/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kiraclaw"
version = "0.1.0"
description = "KiraClaw local daemon and product scaffolding"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "apscheduler>=3.10.4",
    "fastapi>=0.119.0",
    "uvicorn>=0.38.0",
    "pydantic-settings>=2.0.0",
    "krim>=0.4.2",
    "krim-sdk>=0.4.0",
    "slack-bolt>=1.23.0",
    "aiohttp>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
kiraclaw-agentd = "kiraclaw_agentd.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["apps/agentd/src/kiraclaw_agentd"]

[tool.hatch.build.targets.sdist]
include = [
    "/README.md",
    "/pyproject.toml",
    "/apps/agentd/src/kiraclaw_agentd",
    "/docs",
]
